updated libstackmate ios folder and added ffi code for ios
